pyproject.toml
[project]
name = "bot-binance-12"
version = "0.1.0"
description = "Bot de trading para Binance Spot — primero validar, después operar"
requires-python = ">=3.11"
dependencies = [
    "ccxt>=4.4",
    "pandas>=2.2",
    "numpy>=1.26",
    "pandas-ta>=0.3.14b",
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "rich>=13.7",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
asyncio_mode = "auto"
